Revert "perf: (migration): Remove useTable hook from migration farm tables (#4322)"
This reverts commit 600e1df38089ee7d3ff5b955606053dbfa8d6e23.
